refactor: put assignment solver files in own mod
